requirements.txt
discord.py>=2.6.0
aiohttp
torch
torchvision
clip @ git+https://github.com/openai/CLIP.git
Pillow
